+++ module/category/src/Infrastructure/Handler/DeleteCategoryCommandHandler.php
@@ -12,6 +12,8 @@ namespace Ergonode\Category\Infrastructure\Handler;
 use Ergonode\Category\Domain\Command\DeleteCategoryCommand;
 use Ergonode\Category\Domain\Entity\Category;
 use Ergonode\Category\Domain\Repository\CategoryRepositoryInterface;
+use Ergonode\Core\Infrastructure\Exception\ExistingRelationshipsException;
+use Ergonode\Core\Infrastructure\Resolver\RelationshipsResolverInterface;
 use Webmozart\Assert\Assert;
 
 /**
@@ -24,11 +26,20 @@ class DeleteCategoryCommandHandler
     private $repository;
 
     /**
-     * @param CategoryRepositoryInterface $repository
+     * @var RelationshipsResolverInterface
      */
-    public function __construct(CategoryRepositoryInterface $repository)
-    {
+    private $relationshipsResolver;
+
+    /**
+     * @param CategoryRepositoryInterface    $repository
+     * @param RelationshipsResolverInterface $relationshipsResolver
+     */
+    public function __construct(
+        CategoryRepositoryInterface $repository,
+        RelationshipsResolverInterface $relationshipsResolver
+    ) {
         $this->repository = $repository;
+        $this->relationshipsResolver = $relationshipsResolver;
     }
 
     /**
@@ -36,10 +47,15 @@ class DeleteCategoryCommandHandler
      *
      * @throws \Exception
      */
-    public function __invoke(DeleteCategoryCommand $command)
+    public function __invoke(DeleteCategoryCommand $command): void
     {
         $category = $this->repository->load($command->getId());
         Assert::isInstanceOf($category, Category::class, sprintf('Can\'t find category with id "%s"', $command->getId()));
+
+        $relationships = $this->relationshipsResolver->resolve($command->getId());
+        if (!$relationships->isEmpty()) {
+            throw new ExistingRelationshipsException($command->getId());
+        }
 
         $this->repository->delete($category);
     }
